Extract common code to reduce duplication
