+++ helpers/getImageManagerImage.js
@@ -1,7 +1,5 @@
 'use strict';
-const utils = require('handlebars-utils');
-const common = require('./lib/common');
-const SafeString = require('handlebars').SafeString;
+const { getObjectStorageImage } = require('./lib/getObjectStorageImage')
 
 const factory = globals => {
     return function(path) {
@@ -11,24 +9,7 @@ const factory = globals => {
 
         const options = arguments[arguments.length - 1];
 
-        if (!utils.isString (path) || common.isValidURL(path)) {
-            throw new TypeError("Invalid path for getImageManagerImage helper");
-        }
-
-        // Return original image if there are no arguments
-        let size = 'original';
-
-        if (Number.isInteger(options.hash['width'])) {
-            if (Number.isInteger(options.hash['height'])) {
-                // Return image of specified size
-                size = `${options.hash['width']}x${options.hash['height']}`
-            } else {
-                // Return image of specified width
-                size = `${options.hash['width']}w`
-            }
-        }
-
-        return new SafeString(`${cdnUrl}/images/stencil/${size}/image-manager/${path}`);
+        return getObjectStorageImage(cdnUrl, 'image-manager', path, options);
     };
 };
