+++ helpers/getImageManagerImageSrcset.js
@@ -1,18 +1,5 @@
 'use strict';
-const utils = require('handlebars-utils');
-const common = require('./lib/common');
-const SafeString = require('handlebars').SafeString;
-
-const srcsets = {
-    '80w': '80w',
-    '160w': '160w',
-    '320w': '320w',
-    '640w': '640w',
-    '960w': '960w',
-    '1280w': '1280w',
-    '1920w': '1920w',
-    '2560w': '2560w',
-};
+const { getObjectStorageImageSrcset } = require('./lib/getObjectStorageImage')
 
 const factory = globals => {
     return function(path) {
@@ -20,13 +7,7 @@ const factory = globals => {
 
         const cdnUrl = siteSettings.cdn_url || '';
 
-        if (!utils.isString (path) || common.isValidURL(path)) {
-            throw new TypeError("Invalid path for getContentImageSrcset helper");
-        }
-
-        return new SafeString(Object.keys(srcsets).map(descriptor => {
-            return ([`${cdnUrl}/images/stencil/${srcsets[descriptor]}/image-manager/${path} ${descriptor}`].join(' '));
-        }).join(', '));
+        return getObjectStorageImageSrcset(cdnUrl, 'image-manager', path);
     };
 };
